Rename namespace PathMatch to PathMatcher.
